requirements.txt
discord.py==1.7.3
numpy==1.22.1
pandas==1.4.0
scikit_learn==1.0.2
tensorboardX==2.4.1
torch==1.10.2
tqdm==4.62.3
transformers==4.16.2
